Wallpaper & Background Worker
The app is now created on a background thread so that the UI doesn't
freeze up. It also displays a progress bar to show the app's progress as
it is being created. This is helpful for slower machines.

I also added in the ability to add wallpapers to either Windows Phone or
Windows 8.